requirements.txt
prometheus_client
jira
python-dotenv
loguru